feat: move ff and rnn networks into a folder
